requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.36
pydantic==2.10.4
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
